refactor: implement high-precision decimal math and resilient price feed
- Migrated all financial logic from f64 to rust_decimal for deterministic PnL/liquidation
- Refactored EngineEvent to use Decimal types across MPSC channels
- Implemented resilient WebSocket loop with non-blocking reconnection in ws.rs

+++ perp-engine/src/engine/position.rs
@@ -1,21 +1,22 @@
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
+use rust_decimal::Decimal;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub enum PositionType {
     Long,
     Short,
 }
-#[derive(Debug, Serialize, Deserialize, Clone)]
 
+#[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Position {
     pub id: Uuid,
     pub asset: String,
-    pub entry_price: f64,
-    pub quantity: f64,
-    pub margin: f64,
-    pub leverage: f64,
-    pub pnl: f64,
+    pub entry_price: Decimal, 
+    pub quantity: Decimal,    
+    pub margin: Decimal,      
+    pub leverage: Decimal,    
+    pub pnl: Decimal,         
     pub position_type: PositionType,
-    pub liquidation_price: f64,
+    pub liquidation_price: Decimal, 
 }
