+++ perp-engine/src/engine/trade.rs
@@ -1,10 +1,11 @@
 use serde::{Serialize, Deserialize};
 use crate::engine::position::PositionType;
+use rust_decimal::Decimal; 
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct Trade {
-    pub entry: f64,
-    pub exit: f64,
-    pub pnl: f64,
+    pub entry: Decimal,         
+    pub exit: Decimal,          
+    pub pnl: Decimal,           
     pub position_type: PositionType,
 }
